+++ common/src/main/java/dev/boxadactle/coordinatesdisplay/CoordinatesDisplay.java
@@ -44,6 +44,8 @@ public class CoordinatesDisplay {
 
 	public static Vec3<Integer> MARK_POS = null;
 
+	private static boolean boxHudInstalled;
+
 	static {
 		LOGGER.info("Initializing " + MOD_NAME + " v" + VERSION);
 
@@ -96,6 +98,17 @@ public class CoordinatesDisplay {
 			CoordinatesDisplay.CONFIG.resetConfig();
 
 			deltaError = true;
+		}
+	}
+
+	public static boolean isBoxhudInstalled() {
+		try {
+			Class.forName("dev.boxadactle.boxhud.Boxhud");
+			CoordinatesDisplay.LOGGER.info("BoxHud is installed!");
+			return true;
+		} catch (ClassNotFoundException e) {
+			CoordinatesDisplay.LOGGER.info("BoxHud is not installed :(.");
+			return false;
 		}
 	}