+++ common/src/main/java/dev/boxadactle/coordinatesdisplay/Hud.java
@@ -50,6 +50,7 @@ public class Hud {
         bl &= !ClientUtils.getClient().debugEntries.isOverlayVisible();
         bl &= CoordinatesDisplay.shouldHudRender;
         bl &= filter.getFilter().isVisible();
+        bl &= !CoordinatesDisplay.isBoxhudInstalled();
 
         return bl && CoordinatesDisplay.getConfig().enabled;
     }