+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - Added Bedrock-Edition rendering mode
 - Allowing to change size of the 3d compass
 - Fixed bug where HUD would not render over chunk borders
+- Disallow moving HUD when BoxHud is installed